Transparency, debugging, optimization (or not) and more
Lines have an opacity depending on the distance of the two points - the longer the line, the more transparent.
Added debug particles - if the flag is on, there will always be one on the screen. It has a log method which can be used for information. The point was to avoid the log spam when all particles start logging.
Lines are no longer drawn twice - added a check if a line is already added for drawing - was adding every line twice (from the perspective of each of the two particles).
Turned utils to a module.
Default value of DrawParticles is now false.